# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 공통기능 25개 API
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/data-scopes:
    get:
      operationId: listDataScopes
      summary: 데이터 범위 권한 조회
      x-related-requirements:
      - REQ-028
      - REQ-029
      - REQ-030
      - REQ-031
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-144
      - REQ-166
      - REQ-167
      - REQ-176
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveDataScopes
      summary: 데이터 범위 권한 저장/실행
      x-related-requirements:
      - REQ-028
      - REQ-029
      - REQ-030
      - REQ-031
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-144
      - REQ-166
      - REQ-167
      - REQ-176
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      x-business-rules:
      - 데이터 범위 권한 원문 제약과 서버 권한을 적용한다
      - data_scope_permissions 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - data_scope_permissions 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 데이터 범위 권한 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 데이터 범위 권한 저장 -> data_scope_permissions 후속 조회에 반영'
      - 'auth: R01이 데이터 범위 권한 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> data_scope_permissions 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 데이터 범위 권한 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/code-details:
    get:
      operationId: listCodeDetails
      summary: 상세코드 관리 조회
      x-related-requirements:
      - REQ-039
      - REQ-040
      - REQ-041
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-121
      - REQ-122
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-147
      - REQ-166
      - REQ-167
      - REQ-179
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveCodeDetails
      summary: 상세코드 관리 저장/실행
      x-related-requirements:
      - REQ-039
      - REQ-040
      - REQ-041
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-120
      - REQ-121
      - REQ-122
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-147
      - REQ-166
      - REQ-167
      - REQ-179
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      x-business-rules:
      - 상세코드 관리 원문 제약과 서버 권한을 적용한다
      - code_details 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - code_details 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 상세코드 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 상세코드 관리 저장 -> code_details 후속 조회에 반영'
      - 'auth: R01이 상세코드 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> code_details 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 상세코드 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/batch-results:
    get:
      operationId: listBatchResults
      summary: 배치 결과 조회·로그 조회
      x-related-requirements:
      - REQ-100
      - REQ-101
      - REQ-102
      - REQ-103
      - REQ-105
      - REQ-113
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-164
      - REQ-166
      - REQ-167
      - REQ-193
      - REQ-194
      - REQ-198
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveBatchResults
      summary: 배치 결과 조회 저장/실행
      x-related-requirements:
      - REQ-100
      - REQ-101
      - REQ-102
      - REQ-103
      - REQ-105
      - REQ-113
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-164
      - REQ-166
      - REQ-167
      - REQ-193
      - REQ-194
      - REQ-198
      x-roles:
      - R09
      x-business-rules:
      - 배치 결과 조회 원문 제약과 서버 권한을 적용한다
      - batch_results 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - batch_results 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 배치 결과 조회 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 배치 결과 조회 저장 -> batch_results 후속 조회에 반영'
      - 'auth: R01이 배치 결과 조회 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> batch_results 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 배치 결과 조회 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/batch-results/{batchResultId}/log:
    get:
      operationId: getBatchLog
      summary: 배치 로그 조회
      x-related-requirements:
      - REQ-100
      - REQ-101
      - REQ-102
      - REQ-103
      - REQ-113
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-126
      - REQ-164
      - REQ-166
      - REQ-167
      - REQ-193
      - REQ-198
      x-roles:
      - R09
      x-business-rules:
      - 배치 로그 조회 권한과 원문 제약을 검증한다
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        meta:
          type: object
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
        meta:
          type: object
...
security:
- sessionCookie: []
...